# flagsentinel-api/flagsentinel-api/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: FlagSentinel API
  version: 1.0.0
  description: API para gestión, evaluación y administración de Feature Flags
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/rules:
    post:
      tags: [ Rules ]
      summary: Crear una regla
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '200':
          description: Regla creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'

    get:
      tags: [ Rules ]
      summary: Obtener todas las reglas
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Lista de reglas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleResponse'
  /api/rules/paged:
    get:
      tags: [ Rules ]
      summary: Obtener reglas paginadas
      description: Devuelve una página de reglas según los parámetros de paginación.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Página de reglas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRuleResponse'
  /api/rules/{id}:
    get:
      tags: [ Rules ]
      summary: Obtener una regla por ID
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Regla encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'

    put:
      tags: [ Rules ]
      summary: Actualizar una regla
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleRequest'
      responses:
        '200':
          description: Regla actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'

    delete:
      tags: [ Rules ]
      summary: Eliminar una regla
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado correctamente

  ############################################
  # RULE OPERATORS (dinámicos)
  ############################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateRuleRequest:
      type: object
      properties:
        attribute:
          type: string
        operator:
          type: string
        value:
          type: string
    UpdateRuleRequest:
      type: object
      properties:
        attribute:
          type: string
        operator:
          type: string
        value:
          type: string
    RuleResponse:
      type: object
      properties:
        id:
          type: integer
        attribute:
          type: string
        operator:
          type: string
        value:
          type: string
...
    PageRuleResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RuleResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...